core/requirements.txt
celery==5.2.6
requests==2.26.0
SQLAlchemy==1.4.29
psycopg2-binary==2.9.3
pika==1.2.0
python-dotenv==0.21.0